+++ src/net_lib/common/packet_id_counter.hpp
@@ -1,4 +1,4 @@
 #pragma once
 
-int next_server_packet_id();
-int next_client_packet_id();
+uint8_t next_server_packet_id();
+uint8_t next_client_packet_id();
